Prune auto-research shortcut replay path
