+++ loopx/capabilities/auto_research/core.py
@@ -7,6 +7,5 @@ from .kernel import (
     LIGHTWEIGHT_AUTO_RESEARCH_HYPOTHESIS_SCHEMA_VERSION,
     LIGHTWEIGHT_AUTO_RESEARCH_RESULT_SCHEMA_VERSION,
     lightweight_hypothesis,
-    run_builtin_lightweight_demo,
     run_lightweight_auto_research,
 )
