+++ loopx/capabilities/auto_research/kernel.py
@@ -1,8 +1,6 @@
 from __future__ import annotations
 
-import importlib.util
 from collections.abc import Callable, Iterable, Mapping
-from pathlib import Path
 from typing import Any
 
 
@@ -191,79 +189,3 @@ def run_lightweight_auto_research(
             "raw_source_bodies_recorded": False,
         },
     }
-
-
-def _load_knn_pack_evaluator(pack_dir: Path) -> Any:
-    evaluator_path = pack_dir / "protected_eval.py"
-    spec = importlib.util.spec_from_file_location("loopx_auto_research_knn_eval", evaluator_path)
-    if spec is None or spec.loader is None:
-        raise ValueError(f"cannot load protected evaluator: {evaluator_path.name}")
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    if not hasattr(module, "evaluate"):
-        raise ValueError("protected evaluator must define evaluate(solution_path, split)")
-    return module
-
-
-def _default_knn_pack_dir() -> Path:
-    return Path(__file__).resolve().parents[3] / "examples" / "auto_research_knn_pack"
-
-
-def run_builtin_lightweight_demo(
-    *,
-    goal_id: str = "loopx-auto-research-lite",
-    pack_dir: str | Path | None = None,
-) -> dict[str, Any]:
-    """Run the smallest real k-NN auto-research loop over the public pack."""
-
-    pack = Path(pack_dir).expanduser() if pack_dir is not None else _default_knn_pack_dir()
-    evaluator = _load_knn_pack_evaluator(pack)
-    candidates = [
-        lightweight_hypothesis(
-            hypothesis_id="hyp_full_sort",
-            todo_id="todo_auto_research_lite_001",
-            claimed_by="research-curator",
-            text="Keep full sorting as the baseline candidate.",
-            candidate_key="full_sort",
-        ),
-        lightweight_hypothesis(
-            hypothesis_id="hyp_partial_selection",
-            todo_id="todo_auto_research_lite_002",
-            claimed_by="evidence-runner",
-            text="Use exact partial selection before full sorting.",
-            candidate_key="partial_selection",
-        ),
-    ]
-
-    def evaluate(hypothesis: dict[str, Any], split: str) -> Mapping[str, Any]:
-        solution_name = (
-            "solution_baseline.py"
-            if hypothesis["candidate_key"] == "full_sort"
-            else "solution_candidate.py"
-        )
-        result = evaluator.evaluate(pack / solution_name, split)
-        return {
-            "metric": result["metric"]["value"],
-            "exact": result["exact"],
-            "protected_scope_clean": result["protected_scope_clean"],
-            "strategy": result["strategy"],
-            "artifact_refs": result.get("artifact_refs", []),
-            "result_source": "knn_pack_protected_eval",
-        }
-
-    payload = run_lightweight_auto_research(
-        goal_id=goal_id,
-        hypotheses=candidates,
-        evaluate=evaluate,
-        baseline=1.0,
-        direction="maximize",
-        max_dev_rounds=2,
-    )
-    payload["result_source"] = "knn_pack_protected_eval"
-    payload["evaluator"] = {
-        "schema_version": "auto_research_lightweight_evaluator_v0",
-        "kind": "public_knn_pack_protected_eval",
-        "pack": "examples/auto_research_knn_pack",
-        "raw_paths_recorded": False,
-    }
-    return payload
